Changed data type for output

+++ src/main/java/lv/javaguru/java1/student_maksims_latkovskis/lesson_1/Addition.java
@@ -17,7 +17,7 @@ public class Addition {
         int SecondVal =scanner.nextInt();
 
         // Performing summation and printing result
-        int SumVal=FirstVal+SecondVal;
+        long SumVal=FirstVal+SecondVal;
         System.out.println("Your result is " + SumVal);
     }
 }
